refactor: convert css to CSS modules
