+++ src/assets/svg/SortingSvg/sortingSvg.module.css
@@ -1,0 +1,4 @@
+.sorting-icon:hover{
+    cursor:pointer;
+    opacity:0.8
+}
