+++ src/сomponents/BookList/BookList.jsx
@@ -1,13 +1,13 @@
-import "./BookList.css";
+import styles from "./BookList.module.css";
 import { useState } from "react";
 
 const BookList = ({ listBooks }) => {
   const [headline, setHeadline] = useState("Лучшие книги");
 
   return (
-    <div className="book-list-wrapper">
+    <div className={styles["book-list-wrapper"]}>
       <h2>{headline}</h2>
-      <div className="book-list"></div>
+      <div className={styles["book-list"]}></div>
     </div>
   );
 };
