+++ src/сomponents/Footer/Footer.jsx
@@ -1,4 +1,4 @@
-import "./footer.css";
+import styles from "./footer.module.css";
 
 const Footer = () => {
   return <footer></footer>;
