+++ src/сomponents/Header/header.module.css
@@ -24,11 +24,6 @@ h1{
     gap:10px;
 }
 
-.sorting-icon:hover{
-    cursor:pointer;
-    opacity:0.8
-}
-
 input{
     width:100%;
     height:35px;
@@ -63,6 +58,7 @@ button{
 button:hover{
     background-color:var(--second-color)
 }
+
 button:active{
     background-color:#1a2077
 }
@@ -74,8 +70,7 @@ button:active{
     margin:none
 }
 
-@media(max-width:560px){
-
+@media(max-width:576px){
     header{
         height:90px;
     }
@@ -100,6 +95,3 @@ button:active{
         right:90px;
     }
 }
-
-
-
